修改chat.js里面飞机角度调整，修改flyingpathVO 22号类型点过滤

+++ phone-web/src/main/java/hust/phone/web/network/handler/MinaServerHandler.java
@@ -626,7 +626,7 @@ public class MinaServerHandler extends IoHandlerAdapter {
 //		String lat = unpack.GPS_LAT+"";
 		short mode =unpack.BASEMODE;
 		long uavDeciveId = packet.SND_DEVICE_ID;
-		float GPS_HDG =unpack.GPS_HDG;
+		int GPS_HDG =unpack.GPS_HDG;
 		
 		String content = ConstantUtils.MSG_TANS_STATUS+":"+lon+","+lat;
 		switch (mode) {
